The plugin repository server's address can be configured #90 Still defaults to 0.0.0.0:1234

+++ src/Tailviewer.PluginRepository/App.cs
@@ -66,7 +66,7 @@ namespace Tailviewer.PluginRepository
 				using (var repo = PluginRepository.Create())
 				{
 					var app = new TApp();
-					return app.Run(filesystem, repo, options);
+					return (int) app.Run(filesystem, repo, options);
 				}
 			}
 		}
